Consolidate wash stop host and wash down functions.
